Stop these being a partial class
It slows down things like ReSharper trying to update analyses
