+++ ICSharpCode.CodeConverter/CSharp/VisualBasicConverter.cs
@@ -1,29 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using ICSharpCode.CodeConverter.Util;
+﻿using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Formatting;
-using Microsoft.CodeAnalysis.Text;
-using ArrayRankSpecifierSyntax = Microsoft.CodeAnalysis.CSharp.Syntax.ArrayRankSpecifierSyntax;
-using ArrayTypeSyntax = Microsoft.CodeAnalysis.CSharp.Syntax.ArrayTypeSyntax;
-using CSharpExtensions = Microsoft.CodeAnalysis.CSharp.CSharpExtensions;
-using ExpressionSyntax = Microsoft.CodeAnalysis.CSharp.Syntax.ExpressionSyntax;
-using SyntaxFactory = Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
-using SyntaxFacts = Microsoft.CodeAnalysis.CSharp.SyntaxFacts;
-using SyntaxKind = Microsoft.CodeAnalysis.CSharp.SyntaxKind;
-using TypeSyntax = Microsoft.CodeAnalysis.CSharp.Syntax.TypeSyntax;
 using VBasic = Microsoft.CodeAnalysis.VisualBasic;
 using VBSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax;
-using static ICSharpCode.CodeConverter.CSharp.SyntaxKindExtensions;
 
 namespace ICSharpCode.CodeConverter.CSharp
 {
-    public partial class VisualBasicConverter
+    public class VisualBasicConverter
     {
         public static async Task<SyntaxNode> ConvertCompilationTree(Document document,
             CSharpCompilation csharpViewOfVbSymbols)
@@ -34,7 +17,7 @@ namespace ICSharpCode.CodeConverter.CSharp
             var root = (VBasic.VisualBasicSyntaxNode)await document.GetSyntaxRootAsync();
 
             var visualBasicSyntaxVisitor = new 
-                NodesVisitor(compilation, semanticModel, csharpViewOfVbSymbols);
+                DeclarationAndExpressionNodeVisitor(compilation, semanticModel, csharpViewOfVbSymbols);
             return root.Accept(visualBasicSyntaxVisitor.TriviaConvertingVisitor);
         }
     }
